requirements.txt
docopt==0.6.2
numpy==1.15.2
biopython==1.72
pandas==0.23.4
schema==0.6.8
tqdm==4.28.1
wget==3.2
